requirements.txt
torch
torchvision
matplotlib
numpy
Pillow
skimage
tqdm
tensorboard
lpips
